+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This page reports the main releases only and the main changes therein.
 
-## [v1.10.0] - 2025-11-18  
+## [v1.10.0] - 2025-11-19
 **Enhanced noise detection and interface improvements**
 
 ### Added